update: basic site routing for the assignments.
Due to the update in the way it compiles and saves the pdfs,
now I have to recreate that layout in the (html) site.

So, in eleventy.config.js I add a collection that basically just scans the
pdfs folder that will be previously created by the github action and the
it simple returns an array of objects with properties {title, course, url}.

Then those properties are used in the index html that 11ty then process
with liquid in order to give a simple structure (a list) for the assignments.